services/orchestration/pyproject.toml
[tool.poetry]
name = "orchestration-service"
version = "0.1.0"
description = "Airflow orchestration for domain ingestion pipeline"
authors = ["Szymon Wiedeheft"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = ">=3.12,<3.13"
apache-airflow = "^2.10.0"
pydantic = "^2.5.0"
docker = "^6.1.0"
kafka-python = "^2.0.2"
requests = "^2.31.0"
structlog = "^23.2.0"

# Internal dependencies
common = {path = "../../packages/common", develop = true}
schemas = {path = "../../packages/schemas", develop = true}
monitoring = {path = "../../packages/monitoring", develop = true}

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.3"
pytest-cov = "^4.1.0"
black = "^23.12.0"
ruff = "^0.1.9"
mypy = "^1.7.1"

[tool.pytest.ini_options]
pythonpath = [".", "dags", "plugins"]

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
